third_party/WebKit/LayoutTests/fast/css/computed-width-without-renderer.html

<html>
<head>
<script src="../../resources/js-test.js"></script>
<script>
    if (window.testRunner)
        testRunner.dumpAsText();

    function measure() {
        var div = document.createElement('div');
        div.textContent = "PASS";

        var link = document.createElement('link');
        link.rel = 'stylesheet';
        link.href = "doesnotexist.css";
        document.head.appendChild(link);

        document.body.appendChild(div);
        var style = getComputedStyle(div);
        window.width1 = style.width;
        window.width2 = style.width;
        shouldBe("window.width1", "window.width2");
    }
</script>
</head>
<body onload="measure()">
</body>
</html>
